+++ Vector2.h
@@ -382,7 +382,7 @@ inline bool operator !=(const Vector2<T>& left, const Vector2<T>& right)
 // Define the most common types
 typedef Vector2<int>          Vector2i;
 typedef Vector2<unsigned int> Vector2u;
-typedef Vector2<float>        Vector2f;
+typedef Vector2<double>       Vector2f;
 
 
 ////////////////////////////////////////////////////////////
